+++ public/towards/code/src/SugarScape/Model.hs
@@ -35,9 +35,9 @@ import SugarScape.Discrete
 ------------------------------------------------------------------------------------------------------------------------
 data SugAgentState = SugAgentState 
   { sugAgCoord            :: Discrete2dCoord
-  , sugAgSugarMetab       :: Double              -- this amount of sugar will be consumed by the agent in each time-step
+  , sugAgSugarMetab       :: Int              -- this amount of sugar will be consumed by the agent in each time-step
   , sugAgVision           :: Int                 -- the vision of the agent: strongly depends on the type of the environment: Int because its 2d discrete
-  , sugAgSugarLevel       :: Double              -- the current sugar holdings of the agent, if 0 then the agent starves to death
+  , sugAgSugarLevel       :: Int              -- the current sugar holdings of the agent, if 0 then the agent starves to death
     } deriving (Show, Eq)
 
 data SugAgentObservable = SugAgentObservable
@@ -47,12 +47,12 @@ data SugAgentObservable = SugAgentObservable
 
 data SugEnvCellOccupier = SugEnvCellOccupier 
   { sugEnvOccId     :: AgentId
-  , sugEnvOccWealth :: Double
+  , sugEnvOccWealth :: Int
   } deriving (Show, Eq)
 
 data SugEnvCell = SugEnvCell 
-  { sugEnvSugarCapacity :: Double
-  , sugEnvSugarLevel    :: Double
+  { sugEnvSugarCapacity :: Int
+  , sugEnvSugarLevel    :: Int
   , sugEnvOccupier      :: Maybe SugEnvCellOccupier
   } deriving (Show, Eq)
 
@@ -70,14 +70,14 @@ type SugAgentOut g  = AgentOut (SugAgentMonad g) SugAgentObservable
 ------------------------------------------------------------------------------------------------------------------------
 -- SUGARSCAPE PARAMETERS
 ------------------------------------------------------------------------------------------------------------------------
-maxSugarCapacityCell :: Double
-maxSugarCapacityCell = 4.0
+maxSugarCapacityCell :: Int
+maxSugarCapacityCell = 4
 
 data SugarScapeParams = SugarScapeParams 
   { sgAgentCount           :: Int
-  , spSugarGrowBackRate    :: Double            -- negative value means G_inf: regrow to max in next step
-  , spSugarEndowmentRange  :: (Double, Double)
-  , spSugarMetabolismRange :: (Double, Double)
+  , spSugarGrowBackRate    :: Int            -- negative value means G_inf: regrow to max in next step
+  , spSugarEndowmentRange  :: (Int, Int)
+  , spSugarMetabolismRange :: (Int, Int)
   , spVisionRange          :: (Int, Int)
   }
 
@@ -85,8 +85,8 @@ mkSugarScapeParams :: SugarScapeParams
 mkSugarScapeParams = SugarScapeParams {
     sgAgentCount           = 0
   , spSugarGrowBackRate    = 0
-  , spSugarEndowmentRange  = (0.0, 0.0)
-  , spSugarMetabolismRange = (0.0, 0.0)
+  , spSugarEndowmentRange  = (0, 0)
+  , spSugarMetabolismRange = (0, 0)
   , spVisionRange          = (0, 0) 
   }
 
@@ -96,27 +96,27 @@ mkSugarScapeParams = SugarScapeParams {
 mkParamsAnimationII_1 :: SugarScapeParams
 mkParamsAnimationII_1 = mkSugarScapeParams {
     sgAgentCount           = 400         -- page 28
-  , spSugarGrowBackRate    = -1.0        -- regrow to max immediately
-  , spSugarEndowmentRange  = (5.0, 25.0) -- NOTE: this is specified in book page 33 where the initial endowments are set to 5-25
-  , spSugarMetabolismRange = (1.0, 4.0)  -- NOTE: specified where? 1 - 4
+  , spSugarGrowBackRate    = -1        -- regrow to max immediately
+  , spSugarEndowmentRange  = (5, 25) -- NOTE: this is specified in book page 33 where the initial endowments are set to 5-25
+  , spSugarMetabolismRange = (1, 4)  -- NOTE: specified where? 1 - 4
   , spVisionRange          = (1, 6)      -- NOTE: set to 1-6 on page 24
   }
 
 mkParamsAnimationII_2 :: SugarScapeParams
 mkParamsAnimationII_2 = mkSugarScapeParams {
     sgAgentCount           = 400         -- page 28
-  , spSugarGrowBackRate    = 1.0         -- regrow by 1 unit per step
-  , spSugarEndowmentRange  = (5.0, 25.0) -- NOTE: this is specified in book page 33 where the initial endowments are set to 5-25
-  , spSugarMetabolismRange = (1.0, 4.0)  -- NOTE: specified where? 1 - 4
+  , spSugarGrowBackRate    = 1         -- regrow by 1 unit per step
+  , spSugarEndowmentRange  = (5, 25) -- NOTE: this is specified in book page 33 where the initial endowments are set to 5-25
+  , spSugarMetabolismRange = (1, 4)  -- NOTE: specified where? 1 - 4
   , spVisionRange          = (1, 6)      -- NOTE: set to 1-6 on page 24
   }
 
 mkParamsCarryingCapacity :: SugarScapeParams
 mkParamsCarryingCapacity = mkSugarScapeParams {
     sgAgentCount           = 400         -- 400 agents only
-  , spSugarGrowBackRate    = 1.0         -- regrow by 1 unit per step
-  , spSugarEndowmentRange  = (5.0, 25.0) -- NOTE: this is specified in book page 33 where the initial endowments are set to 5-25
-  , spSugarMetabolismRange = (1.0, 4.0)  -- NOTE: specified where? 1 - 4
+  , spSugarGrowBackRate    = 1         -- regrow by 1 unit per step
+  , spSugarEndowmentRange  = (5, 25) -- NOTE: this is specified in book page 33 where the initial endowments are set to 5-25
+  , spSugarMetabolismRange = (1, 4)  -- NOTE: specified where? 1 - 4
   , spVisionRange          = (1, 6)      -- NOTE: set to 1-6 on page 24
   }
